finished week 4 team activity
