+++ src/cart/index.html
@@ -11,6 +11,7 @@
     <link rel="stylesheet" href="/css/style.css" />
 
     <script src="../js/cart.js" type="module"></script>
+    <script type="module" src="../js/shoppingCart.mjs"></script>
   </head>
 
   <body>
